# Conflicts:
#	LOA_lab3/src/util/GameState.java

+++ LOA_lab3/src/Algo/Heuristic/Concentration.java
@@ -6,7 +6,7 @@ import util.Pos2D;
 public class Concentration implements HeuristicInterface{
 
 	@Override
-	public int getScore(GameState gameState) {
+	public float getScore(GameState gameState) {
 		return 0;
 	}
